requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.9.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT>=2.9.0
mysql-connector-python==8.3.0
pymongo==4.6.1
redis==5.0.1
google-genai>=1.67.0
httpx>=0.28.0
boto3>=1.35.0
aws-sdk-bedrock-runtime>=0.4.0
smithy-core>=0.1.0
openai>=1.0.0
email-validator>=2.0.0
